backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# PDF generation
reportlab==4.2.5

# File upload support
python-multipart==0.0.12

# Rate limiting
slowapi==0.1.9

# Dev / testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0
